Рефакторю код
