+++ js/main.js
@@ -3,7 +3,6 @@ import { renderCards } from './render.js';
 import { getData } from './api.js';
 import { showAlert } from './notifications.js';
 import { initFilters } from './filter.js';
-import './image-preview.js';
 
 getData()
   .then((cards) => {
